+++ core/list_formula_data.php
@@ -28,7 +28,7 @@ if($s != ''){
    $f = "WHERE 1 AND (name LIKE '%".$s."%' OR product_name LIKE '%".$s."%' OR notes LIKE '%".$s."%')";
 }
 
-$formulas = mysqli_query($conn, "SELECT id,fid,name,product_name,isProtected,profile,sex,created,catClass,isMade,madeOn FROM formulasMetaData $f $extra LIMIT $row, $limit");
+$formulas = mysqli_query($conn, "SELECT id,fid,name,product_name,isProtected,profile,sex,created,catClass,isMade,madeOn,status FROM formulasMetaData $f $extra LIMIT $row, $limit");
 
 while ($allFormulas = mysqli_fetch_array($formulas)){
 	    $formula[] = $allFormulas;
@@ -46,6 +46,7 @@ foreach ($formula as $formula) {
 	$r['ingredients'] = (int)countElement("formulas WHERE fid = '".$formula['fid']."'",$conn)?:'0';
 	$r['isMade'] = (int)$formula['isMade']?:0;
 	$r['madeOn'] = (string)$formula['madeOn']?:'N/A';
+	$r['status'] = (int)$formula['status']?:'0';
 
 	$rx[]=$r;