+++ db/pvault.sql
@@ -98,7 +98,8 @@ CREATE TABLE `formulasMetaData` (
   `finalType` INT NOT NULL DEFAULT '100',
   `isMade` INT NOT NULL DEFAULT '0',
   `madeOn` DATETIME NULL DEFAULT NULL,
-  `customer_id` INT NOT NULL DEFAULT '0'
+  `customer_id` INT NOT NULL DEFAULT '0',
+  `status` INT NOT NULL DEFAULT '0'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 DROP TABLE IF EXISTS `IFRALibrary`;