+++ db/updates/update_4.3-4.4.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `formulasMetaData` ADD `status` INT NOT NULL DEFAULT '0' AFTER `customer_id`;