+++ pages/getFormMeta.php
@@ -142,7 +142,20 @@ while($customers = mysqli_fetch_array($cust)){
     </td>
   </tr>
   <tr>
-    <td>Sex:</td>
+    <td>Status:</td>
+    <td>
+    <select name="status" id="status" class="form-control ellipsis">  
+        <option value="0" <?php if($info['status'] == "0"){ echo 'selected';}?>>Schedulled</option>
+        <option value="1" <?php if($info['status'] == "1"){ echo 'selected';}?>>Under Developent</option>
+        <option value="2" <?php if($info['status'] == "2"){ echo 'selected';}?>>Under Evaluation</option>
+        <option value="3" <?php if($info['status'] == "3"){ echo 'selected';}?>>In Production</option>
+        <option value="4" <?php if($info['status'] == "4"){ echo 'selected';}?>>To be reformulated</option>
+        <option value="5" <?php if($info['status'] == "5"){ echo 'selected';}?>>Failure</option>
+    </select>    
+    </td>
+  </tr>
+  <tr>
+    <td>Gender:</td>
     <td><a href="#" id="sex" data-type="select" data-pk="<?php echo $info['id'];?>" data-title="Select sex"></a></td>
   </tr>
   <tr>
@@ -291,6 +304,23 @@ $("#finalType").change(function() {
 	}
   });
 });
+
+$("#status").change(function() {
+ $.ajax({ 
+	url: 'update_data.php', 
+	type: 'GET',
+	data: {
+		formula: '<?=$info['id']?>',
+		updateStatus: 1,
+		formulaStatus: $("#status").find(":selected").val(),
+		},
+	dataType: 'html',
+	success: function (data) {
+		$('#msg').html(data);
+	}
+  });
+});
+
 
 $("#customer").change(function() {
  $.ajax({