+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add formula status (development, failed, production, reformulation)
 - Improve ingredient adding in formula
 - Added a count the total number of suppliers and documents next to each ingredient
+- Add formula name in formula page title
 
 ### Version 4.3
 - Get ingredients list in formula from a json file