+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Improve ingredient adding in formula
 - Added a count the total number of suppliers and documents next to each ingredient
 - Add formula name in formula page title
+- Fix a bug preventing showing the solvent when formula is locked
 
 ### Version 4.3
 - Get ingredients list in formula from a json file